Add package termios
Package termios provides a platform-dependent method to set and restore
raw mode. Ultimately, it may not be needed, but there are subtle
differences in how golang.org/x/term enables raw mode and the method
described in Snaptoken's kilo guide, so I'm keeping it until I'm certain
that it isn't required.