Refactor editor clear behaviour
To minimize the number of flushes from the output buffer to the OS, I've
removed the call to flush from (*editor).clearScreen and created a
public (*editor).Clear function that calls both clearScreen and flush,
to be used when shutting down the editor. This allows us to call flush
only once per iteration of the main render loop.