Calculate line length based on rune count
To better limit the cursor position to the end of a given line, we
must know the number of runes in the string. Switching to a rune-based
representation of lines greatly simplifies this at the expense of
copying the line to and from its string representation when read from
input or written to output.

This implementation doesn't yet handle multi-rune grapheme clusters,
which may break in unexpected lines, or cause the cursor to be
incorrectly positioned relative to the character perceived by the user.