Replace bytes.Buffer with bufio.Writer
On further inspection of the bufio source code, bufio.Writer flushes to
output only if its internal buffer is not long enough to hold the
entirety of the bytes being written to it. This is appropriate for our
use case, so rather the reinvent the wheel, the TerminalWriter's
bytes.Buffer and io.Writer have been replaced by a bytes.Buffer
instance.